test: add integration test for "compress_selectors"
